refactor: pages directory structur
