+++ src/pages/index.ts
@@ -1,1 +1,3 @@
-export * from './home/Home';
+export * from './character';
+export * from './home';
+export * from './notFound';
