+++ src/pages/notFound/notFound.tsx
@@ -1,0 +1,3 @@
+export function notFound() {
+  return <div>notFound</div>;
+}
